# Expand performance tests comparison & output