--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -65,9 +65,18 @@ jobs:
   # - Install MU plugin.
   # - Run performance tests (current commit).
   # - Print performance tests results.
+  # - Check out target commit (target branch or previous commit).
+  # - Install npm dependencies.
+  # - Build WordPress.
+  # - Run performance tests (previous/target commit).
+  # - Print target performance tests results.
   # - Set the environment to the baseline version.
   # - Run baseline performance tests.
-  # - Print base line performance tests results.
+  # - Print baseline performance tests results.
+  # - Compare results with base.
+  # - Check if a comment was already made
+  # - Comment on PR with test results
+  # - Add workflow summary
   # - Set the base sha.
   # - Set commit details.
   # - Publish performance results.
@@ -152,7 +161,32 @@ jobs:
         run: npm run test:performance
 
       - name: Print performance tests results
-        run: "node ./tests/performance/results.js"
+        run: node ./tests/performance/results.js
+
+      - name: Check out target commit (target branch or previous commit)
+        run: |
+          if [[ -z "$BASE_REF" ]]; then
+          git fetch -n origin $BASE_SHA
+          git reset --hard $BASE_SHA
+          else
+          git fetch -n origin $BASE_REF
+          git reset --hard $BASE_SHA
+          fi
+        env:
+        BASE_REF: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.ref || '' }}
+        BASE_SHA: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.sha || github.event.before }}
+
+      - name: Install npm dependencies
+        run: npm ci
+
+      - name: Build WordPress
+        run: npm run build
+
+      - name: Run target performance tests (base/previous commit)
+        run: npm run test:performance -- --prefix=before
+
+      - name: Print target performance tests results
+        run: node ./tests/performance/results.js --prefix=before
 
       - name: Set the environment to the baseline version
         run: |
@@ -162,8 +196,30 @@ jobs:
       - name: Run baseline performance tests
         run: npm run test:performance -- --prefix=base
 
-      - name: Print base line performance tests results
-        run: "node ./tests/performance/results.js --prefix=base"
+      - name: Print baseline performance tests results
+        run: node ./tests/performance/results.js --prefix=base
+
+      - name: Compare results with base
+        run: node ./tests/performance/compare-results.js ${{ runner.temp }}/summary.md
+
+      - name: Check if a comment was already made
+        id: find-comment
+        uses: peter-evans/find-comment@a54c31d7fa095754bfef525c0c8e5e5674c4b4b1
+        with:
+        issue-number: ${{ github.event.pull_request.number }}
+        body-includes: Performance test results for
+
+      - name: Comment on PR with test results
+        uses: peter-evans/create-or-update-comment@c6c9a1a66007646a28c153e2a8580a5bad27bcfa
+        with:
+        issue-number: ${{ github.event.pull_request.number }}
+        comment-id: ${{ steps.find-comment.outputs.comment-id }}
+        edit-mode: replace
+        body-path: ${{ runner.temp }}/summary.md
+        token: ${{ secrets.PR_COMMENT_TOKEN }}
+
+      - name: Add workflow summary
+        run: cat ${{ runner.temp }}/summary.md >> $GITHUB_STEP_SUMMARY
 
       - name: Set the base sha
         # Only needed when publishing results.