--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -173,8 +173,8 @@ jobs:
           git reset --hard $BASE_SHA
           fi
         env:
-        BASE_REF: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.ref || '' }}
-        BASE_SHA: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.sha || github.event.before }}
+          BASE_REF: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.ref || '' }}
+          BASE_SHA: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.sha || github.event.before }}
 
       - name: Install npm dependencies
         run: npm ci
@@ -206,17 +206,17 @@ jobs:
         id: find-comment
         uses: peter-evans/find-comment@a54c31d7fa095754bfef525c0c8e5e5674c4b4b1
         with:
-        issue-number: ${{ github.event.pull_request.number }}
-        body-includes: Performance test results for
+          issue-number: ${{ github.event.pull_request.number }}
+          body-includes: Performance test results for
 
       - name: Comment on PR with test results
         uses: peter-evans/create-or-update-comment@c6c9a1a66007646a28c153e2a8580a5bad27bcfa
         with:
-        issue-number: ${{ github.event.pull_request.number }}
-        comment-id: ${{ steps.find-comment.outputs.comment-id }}
-        edit-mode: replace
-        body-path: ${{ runner.temp }}/summary.md
-        token: ${{ secrets.PR_COMMENT_TOKEN }}
+          issue-number: ${{ github.event.pull_request.number }}
+          comment-id: ${{ steps.find-comment.outputs.comment-id }}
+          edit-mode: replace
+          body-path: ${{ runner.temp }}/summary.md
+          token: ${{ secrets.PR_COMMENT_TOKEN }}
 
       - name: Add workflow summary
         run: cat ${{ runner.temp }}/summary.md >> $GITHUB_STEP_SUMMARY