--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -167,11 +167,11 @@ jobs:
       - name: Check out target commit (target branch or previous commit)
         run: |
           if [[ -z "$TARGET_REF" ]]; then
-          git fetch -n origin $TARGET_SHA
-          git reset --hard $TARGET_SHA
+            git fetch -n origin $TARGET_SHA
+            git reset --hard $TARGET_SHA
           else
-          git fetch -n origin $TARGET_REF
-          git reset --hard $TARGET_SHA
+            git fetch -n origin $TARGET_REF
+            git reset --hard $TARGET_SHA
           fi
         env:
           TARGET_REF: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.ref || '' }}