--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -168,11 +168,10 @@ jobs:
         run: |
           if [[ -z "$TARGET_REF" ]]; then
             git fetch -n origin $TARGET_SHA
-            git reset --hard $TARGET_SHA
           else
             git fetch -n origin $TARGET_REF
-            git reset --hard $TARGET_SHA
           fi
+          git reset --hard $TARGET_SHA
         env:
           TARGET_REF: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.ref || '' }}
           TARGET_SHA: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.sha || github.event.before }}