@swissspidy 2023-08-15
**Performance Test Results**

Performance test results for 3f66e3126d26302b073b0e56039c096961dbdb21 are in 🛎️!

This compares the results from this commit with the ones from f6dde69ea39693a6735b89a90ab657a83d7cb1c4.

**Home block theme**

| Metric | Before | After | Diff abs. | Diff % |
| ------ | ------ | ------ | ------ | ------ |
| wpBeforeTemplate | 32.66 ms | 32.98 ms | 0.32 ms | 1.00 % |
| wpTemplate | 45.94 ms | 46.60 ms | 0.66 ms | 1.00 % |
| wpTotal | 79.19 ms | 78.85 ms | -0.34 ms | 0.00 % |

**Home classic theme**

| Metric | Before | After | Diff abs. | Diff % |
| ------ | ------ | ------ | ------ | ------ |
| wpBeforeTemplate | 23.24 ms | 22.83 ms | -0.41 ms | -2.00 % |
| wpTemplate | 57.67 ms | 51.78 ms | -5.90 ms | -11.00 % |
| wpTotal | 80.97 ms | 75.88 ms | -5.09 ms | -7.00 % |